Refactor economy package structure.
Renamed and updated references for the "economyrevamp" package to "economy" for better clarity and consistency across the project. This includes file paths, imports, and related classes. No functional changes were made.
