+++ quickshop-api/src/main/java/com/ghostchu/quickshop/api/economy/benefit/BenefitsAlreadyException.java
@@ -1,4 +1,4 @@
-package com.ghostchu.quickshop.api.economyrevamp.benefit;
+package com.ghostchu.quickshop.api.economy.benefit;
 /*
  * QuickShop-Hikari
  * Copyright (C) 2025 Daniel "creatorfromhell" Vidmar
@@ -17,25 +17,11 @@ package com.ghostchu.quickshop.api.economyrevamp.benefit;
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
 
-import java.math.BigDecimal;
-
 /**
- * BenefitOverflowException
+ * BenefitsAlreadyException
  *
  * @author creatorfromhell
  * @since 6.2.0.11
  */
-public class BenefitOverflowException extends Exception {
-
-  private final BigDecimal overflow;
-
-  public BenefitOverflowException(final BigDecimal overflow) {
-
-    this.overflow = overflow;
-  }
-
-  public BigDecimal benefit() {
-
-    return overflow;
-  }
+public class BenefitsAlreadyException extends Exception {
 }
