+++ quickshop-api/src/main/java/com/ghostchu/quickshop/api/economy/transaction/TransactionCallback.java
@@ -1,4 +1,4 @@
-package com.ghostchu.quickshop.economyrevamp.transaction;
+package com.ghostchu.quickshop.api.economy.transaction;
 /*
  * QuickShop-Hikari
  * Copyright (C) 2025 Daniel "creatorfromhell" Vidmar
@@ -17,19 +17,15 @@ package com.ghostchu.quickshop.economyrevamp.transaction;
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
 
-import com.ghostchu.quickshop.api.economyrevamp.transaction.TransactionCallback;
-import com.ghostchu.quickshop.util.logger.Log;
 import org.jetbrains.annotations.NotNull;
 
-import java.util.logging.Level;
-
 /**
- * QSTransactionCallback
+ * TransactionCallback
  *
  * @author creatorfromhell
  * @since 6.2.0.11
  */
-public interface QSTransactionCallback extends TransactionCallback {
+public interface TransactionCallback {
 
   /**
    * Calling while Transaction commit
@@ -38,7 +34,7 @@ public interface QSTransactionCallback extends TransactionCallback {
    *
    * @return Does commit event has been cancelled
    */
-  default boolean onCommit(@NotNull final QSEconomyTransaction economyTransaction) {
+  default boolean onCommit(@NotNull final EconomyTransaction economyTransaction) {
 
     return true;
   }
@@ -49,9 +45,8 @@ public interface QSTransactionCallback extends TransactionCallback {
    *
    * @param economyTransaction Transaction
    */
-  default void onFailed(@NotNull final QSEconomyTransaction economyTransaction) {
+  default void onFailed(@NotNull final EconomyTransaction economyTransaction) {
 
-    Log.transaction(Level.WARNING, "Transaction failed: " + economyTransaction.lastError() + ", transaction: " + economyTransaction);
   }
 
   /**
@@ -59,9 +54,8 @@ public interface QSTransactionCallback extends TransactionCallback {
    *
    * @param economyTransaction Transaction
    */
-  default void onSuccess(@NotNull final QSEconomyTransaction economyTransaction) {
+  default void onSuccess(@NotNull final EconomyTransaction economyTransaction) {
 
-    Log.transaction("Transaction succeed: " + economyTransaction);
   }
 
   /**
@@ -70,8 +64,8 @@ public interface QSTransactionCallback extends TransactionCallback {
    *
    * @param economyTransaction Transaction
    */
-  default void onTaxFailed(@NotNull final QSEconomyTransaction economyTransaction) {
+  default void onTaxFailed(@NotNull final EconomyTransaction economyTransaction) {
 
-    Log.transaction(Level.WARNING, "Tax Transaction failed: " + economyTransaction.lastError() + ", transaction: " + economyTransaction);
   }
+
 }
