+++ quickshop-bukkit/src/main/java/com/ghostchu/quickshop/QuickShop.java
@@ -11,7 +11,7 @@ import com.ghostchu.quickshop.api.QuickShopProvider;
 import com.ghostchu.quickshop.api.RankLimiter;
 import com.ghostchu.quickshop.api.command.CommandManager;
 import com.ghostchu.quickshop.api.database.DatabaseHelper;
-import com.ghostchu.quickshop.api.economyrevamp.EconomyManager;
+import com.ghostchu.quickshop.api.economy.EconomyManager;
 import com.ghostchu.quickshop.api.event.QSConfigurationReloadEvent;
 import com.ghostchu.quickshop.api.inventory.InventoryWrapperManager;
 import com.ghostchu.quickshop.api.inventory.InventoryWrapperRegistry;
@@ -34,8 +34,8 @@ import com.ghostchu.quickshop.common.util.QuickExecutor;
 import com.ghostchu.quickshop.database.DatabaseIOUtil;
 import com.ghostchu.quickshop.database.HikariUtil;
 import com.ghostchu.quickshop.database.SimpleDatabaseHelperV2;
-import com.ghostchu.quickshop.economyrevamp.EconomyLoader;
-import com.ghostchu.quickshop.economyrevamp.QSEconomyManager;
+import com.ghostchu.quickshop.economy.EconomyLoader;
+import com.ghostchu.quickshop.economy.QSEconomyManager;
 import com.ghostchu.quickshop.listener.BlockListener;
 import com.ghostchu.quickshop.listener.BungeeListener;
 import com.ghostchu.quickshop.listener.ChatListener;
