+++ quickshop-bukkit/src/main/java/com/ghostchu/quickshop/command/subcommand/SubCommand_Benefit.java
@@ -3,9 +3,9 @@ package com.ghostchu.quickshop.command.subcommand;
 import com.ghostchu.quickshop.QuickShop;
 import com.ghostchu.quickshop.api.command.CommandHandler;
 import com.ghostchu.quickshop.api.command.CommandParser;
-import com.ghostchu.quickshop.api.economyrevamp.benefit.BenefitOverflowException;
-import com.ghostchu.quickshop.api.economyrevamp.benefit.BenefitProvider;
-import com.ghostchu.quickshop.api.economyrevamp.benefit.BenefitsAlreadyException;
+import com.ghostchu.quickshop.api.economy.benefit.BenefitOverflowException;
+import com.ghostchu.quickshop.api.economy.benefit.BenefitProvider;
+import com.ghostchu.quickshop.api.economy.benefit.BenefitsAlreadyException;
 import com.ghostchu.quickshop.api.event.Phase;
 import com.ghostchu.quickshop.api.event.settings.type.benefit.ShopBenefitAddEvent;
 import com.ghostchu.quickshop.api.event.settings.type.benefit.ShopBenefitRemoveEvent;
